% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_qcew.R
\name{download_qcew}
\alias{download_qcew}
\title{Download BLS QCEW county high-level zip file for a single year}
\usage{
download_qcew(year, staging_dir = "data/qcew")
}
\arguments{
\item{year}{Integer. Year to download (1990 or later).}

\item{staging_dir}{Character. Directory to save zip files. Default: \code{"data/qcew"}.}
}
\description{
Download BLS QCEW county high-level zip file for a single year
}
\keyword{internal}
