% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qcew.R
\name{pull_annual_pay}
\alias{pull_annual_pay}
\title{Pull average annual pay across multiple years}
\usage{
pull_annual_pay(years, staging_dir = "data/qcew")
}
\arguments{
\item{years}{Integer vector. Years to pull.}

\item{staging_dir}{Character. Local staging directory. Default: \code{"data/qcew"}.}
}
\description{
Pull average annual pay across multiple years
}
\keyword{internal}
