% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qcew.R
\name{pull_employment}
\alias{pull_employment}
\title{Pull total employment across multiple years}
\usage{
pull_employment(years, staging_dir = "data/qcew")
}
\arguments{
\item{years}{Integer vector. Years to pull.}

\item{staging_dir}{Character. Local staging directory. Default: \code{"data/qcew"}.}
}
\description{
Pull total employment across multiple years
}
\keyword{internal}
