% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qcew.R
\name{pull_hhi}
\alias{pull_hhi}
\title{Pull Herfindahl-Hirschman Index of employment concentration}
\usage{
pull_hhi(years, staging_dir = "data/qcew")
}
\arguments{
\item{years}{Integer vector. Years to pull.}

\item{staging_dir}{Character. Local staging directory. Default: \code{"data/qcew"}.}
}
\description{
Pull Herfindahl-Hirschman Index of employment concentration
}
\keyword{internal}
