% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qcew.R
\name{pull_sectoral_pay}
\alias{pull_sectoral_pay}
\title{Pull employment-weighted average annual pay by CORI super-sector}
\usage{
pull_sectoral_pay(years, staging_dir = "data/qcew")
}
\arguments{
\item{years}{Integer vector. Years to pull.}

\item{staging_dir}{Character. Local staging directory. Default: \code{"data/qcew"}.}
}
\description{
Pull employment-weighted average annual pay by CORI super-sector
}
\keyword{internal}
